refactor: move apprise utils to its own module
