+++ cyberdrop_dl/utils/args.py
@@ -12,7 +12,7 @@ from pydantic import BaseModel, Field, ValidationError, computed_field, field_va
 from cyberdrop_dl import __version__
 from cyberdrop_dl.config_definitions import ConfigSettings, GlobalSettings
 from cyberdrop_dl.config_definitions.custom_types import AliasModel, HttpURL
-from cyberdrop_dl.utils.utilities import handle_validation_error
+from cyberdrop_dl.utils.yaml import handle_validation_error
 
 if TYPE_CHECKING:
     from pydantic.fields import FieldInfo
